requirements.txt
flask-cors
flask
openai>=1.40.0
python-dotenv>=1.0.1